requirements.txt
twikit>=1.0.0
pandas>=1.5.0
asyncio
nest-asyncio
openpyxlfastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose
